+++ Apps/Examples/Examples/All Examples/Annotations/AnimatedMarkerExample.swift
@@ -55,7 +55,7 @@ final class AnimatedMarkerExample: UIViewController, ExampleProtocol {
             label.centerXAnchor.constraint(equalTo: view.centerXAnchor)
         ])
 
-        mapView.mapboxMap.loadStyleURI(.satelliteStreets)
+        mapView.mapboxMap.loadStyle(.satelliteStreets)
 
         // add a tap gesture recognizer that will allow the marker to be animated
         mapView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(updatePosition(_:))))